requirements.txt
discord.py
sqlalchemy
toml